// workspaces/aws-cost-insights/plugins/cost-insights-backend/src/service/types.ts
/**
 * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 * Licensed under the Apache License, Version 2.0 (the "License").
 * You may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *   http://www.apache.org/licenses/LICENSE-2.0
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { CompoundEntityRef } from '@backstage/catalog-model';
import { BackstageCredentials } from '@backstage/backend-plugin-api';
import {
  Cost,
  Project,
} from '@backstage-community/plugin-cost-insights-common';
import { Granularity } from '@aws-sdk/client-cost-explorer';

export interface CostInsightsAwsService {
  getCatalogEntityRangeCost(options: {
    entityRef: CompoundEntityRef;
    startDate: Date;
    endDate: Date;
    granularity: Granularity;
    credentials: BackstageCredentials;
  }): Promise<Cost>;

  getCatalogEntityDailyCost(options: {
    entityRef: CompoundEntityRef;
    intervals: string;
    credentials?: BackstageCredentials;
  }): Promise<Cost>;

  listProjects(options: {
    credentials?: BackstageCredentials;
  }): Promise<Project[]>;

  getProjectDailyCost(options: {
    project: string;
    intervals: string;
    credentials?: BackstageCredentials;
  }): Promise<Cost>;

  getOrgDailyCost(options: {
    intervals: string;
    credentials?: BackstageCredentials;
  }): Promise<Cost>;
}